Split projection related code into different files
